+++ usipipo_commons/domain/enums/__init__.py
@@ -1,17 +1,21 @@
-from .vpn_type import VpnType
-from .key_type import KeyType
-from .key_status import KeyStatus
-from .payment_status import PaymentStatus
-from .payment_method import PaymentMethod
+from .billing_status import BillingStatus
+from .consumption_payment_method import ConsumptionPaymentMethod
 from .crypto_order_status import CryptoOrderStatus
-from .crypto_transaction_status import CryptoTransactionStatus
+from .invoice_status import InvoiceStatus
+from .key_status import KeyStatus
+from .payment_method import PaymentMethod
+from .payment_status import PaymentStatus
+from .plan_type import PlanType
+from .vpn_type import VpnType
 
 __all__ = [
     "VpnType",
-    "KeyType",
     "KeyStatus",
     "PaymentStatus",
     "PaymentMethod",
+    "BillingStatus",
+    "InvoiceStatus",
+    "ConsumptionPaymentMethod",
     "CryptoOrderStatus",
-    "CryptoTransactionStatus",
+    "PlanType",
 ]